Refs #496
Adding hex-content to http api to support binary message sending similarly to the rest api, this will avoid the latter to decode the binary message before sending it to http api, and will also provide a long-awaited feature to http api: sending binary messages.

Another requested feature is the ability to use underscores instead of dashes (Anshu's request), this was reported to be a .Net limitation for handling json content.