DLR wait for final_stats in httpapi
In a shipment via SMPP the DLR manager waits for a PDU with one of the final statuses to remove the DLR from redis. In an HTTP submission this did not happen. Thus, DELIVRD were lost if a PDU with an ACCEPTD was sent first even though the HTTP send was made with DLR_LEVEL = 3